adding permission feature

+++ app/controllers/InventoryController.java
@@ -19,6 +19,7 @@ import org.pac4j.core.client.Client;
 import views.html.inventory.*;
 
 import models.*;
+import models.User;
 
 import org.pac4j.play.java.RequiresAuthentication;
 
@@ -27,17 +28,34 @@ public class InventoryController extends UserProfileController<CommonProfile> {
 	@RequiresAuthentication(clientName = "CasClient")
 	@Transactional
 	public Result index(){
-		CommonProfile profile = getUserProfile();
-		return ok(testname.render(profile.getId()));
+		if(checkPrivileges())
+		{
+			flash("Insufficient Privileges");
+			return ok(views.html.index.render());
+		}
+		
+		return ok(index.render());
 	}
 
 	@Transactional
 	public Result show(Integer id) {
+		if(checkPrivilegesAdmin())
+		{
+			flash("Insufficient Privileges");
+			return ok(views.html.index.render());
+		}
+		
 		return ok(show.render(Inventory.findById(id)));
 	}
 
 	@Transactional(readOnly=true)
 	public Result edit(Integer id) {
+		if(checkPrivileges())
+		{
+			flash("Insufficient Privileges");
+			return ok(views.html.index.render());
+		}
+		
 		Form<Inventory> invForm = form(Inventory.class).fill(
 			Inventory.findById(id));
 		return ok(edit.render(id, invForm));
@@ -45,11 +63,23 @@ public class InventoryController extends UserProfileController<CommonProfile> {
 
 	@Transactional(readOnly=true)
 	public Result create() {
+		if(checkPrivilegesAdmin())
+		{
+			flash("Insufficient Privileges");
+			return ok(views.html.index.render());
+		}
+		
 		Form<Inventory> invForm = form(Inventory.class);
 		return ok(create.render(invForm));
 	}
 	@Transactional
 	public Result save() {
+		if(checkPrivileges())
+		{
+			flash("Insufficient Privileges");
+			return ok(views.html.index.render());
+		}
+		
 		Form<Inventory> invForm = form(Inventory.class).bindFromRequest();
 
 		if(invForm.hasErrors()) {
@@ -61,12 +91,24 @@ public class InventoryController extends UserProfileController<CommonProfile> {
 	}
 	@Transactional
 	public Result delete(Integer id) {
+		if(checkPrivilegesAdmin())
+		{
+			flash("Insufficient Privileges");
+			return ok(views.html.index.render());
+		}
+		
 		Inventory.findById(id).delete();
 		flash("success", "Inventory item has been deleted");
 		return ok(index.render());
 	}
 	@Transactional
 	public Result update(Integer id) {
+		if(checkPrivileges())
+		{
+			flash("Insufficient Privileges");
+			return ok(views.html.index.render());
+		}
+		
 		Form<Inventory> invForm = form(Inventory.class).bindFromRequest();
 		if(invForm.hasErrors()) {
 			return badRequest(edit.render(id, invForm));
@@ -75,5 +117,28 @@ public class InventoryController extends UserProfileController<CommonProfile> {
 			flash("success", "Item " + id + " has been updated");
 			return ok(show.render(Inventory.findById(id)));
 		}		
+	}
+	
+	//This function returns false if current user does not possess any roles
+	public boolean checkPrivileges()
+	{
+		CommonProfile profile = getUserProfile();
+		//if you don't have admin role then redirect back to dashboard
+		if((User.findById(profile.getId()).role == null))
+		{
+			return false;
+		}
+		return true;
+	}
+	//This function returns false if current user does not possess the role Admin or SuperAdmin
+	public boolean checkPrivilegesAdmin()
+	{
+		CommonProfile profile = getUserProfile();
+		//if you don't have admin role then redirect back to dashboard
+		if(!(User.findById(profile.getId()).role != User.Role.Admin ||  User.findById(profile.getId()).role != User.Role.SuperAdmin))
+		{
+			return false;
+		}
+		return true;
 	}
 }
