+++ app/controllers/UserController.java
@@ -19,22 +19,46 @@ public class UserController extends UserProfileController<CommonProfile> {
 	@RequiresAuthentication(clientName = "CasClient") 
 	@Transactional
 	public Result index() {
+		if(checkPrivilegesAdmin())
+		{
+			flash("Insufficient Privileges");
+			return ok(views.html.index.render());
+		}
+		
 		return ok(index.render());
 	}
 
 	@Transactional
 	public Result show(String id) {
+		if(checkPrivilegesAdmin())
+		{
+			flash("Insufficient Privileges");
+			return ok(views.html.index.render());
+		}
+		
 		return ok(show.render(User.findById(id)));
 	}
 
 	@Transactional
 	public Result create() {
+		if(checkPrivilegesAdmin())
+		{
+			flash("Insufficient Privileges");
+			return ok(views.html.index.render());
+		}
+		
 		Form<User> userForm = form(User.class);
 		return ok(create.render(userForm));
 	}
 
 	@Transactional
 	public Result delete(String id) {
+		if(checkPrivilegesAdmin())
+		{
+			flash("Insufficient Privileges");
+			return ok(views.html.index.render());
+		}
+		
 		User.findById(id).delete();
 		flash("success", "User item has been deleted");
 		return ok(index.render());
@@ -42,6 +66,12 @@ public class UserController extends UserProfileController<CommonProfile> {
 
 	@Transactional 
 	public Result save() {
+		if(checkPrivileges())
+		{
+			flash("Insufficient Privileges");
+			return ok(views.html.index.render());
+		}
+		
 		Form<User> userForm = form(User.class).bindFromRequest();
 		if(userForm.hasErrors()){
 			return badRequest(create.render(userForm));
@@ -53,6 +83,12 @@ public class UserController extends UserProfileController<CommonProfile> {
 
 	@Transactional(readOnly=true)
 	public Result edit(String id) {
+		if(checkPrivilegesAdmin())
+		{
+			flash("Insufficient Privileges");
+			return ok(views.html.index.render());
+		}
+		
 		Form<User> userForm = form(User.class).fill(
 			User.findById(id));
 		return ok(edit.render(id, userForm));
@@ -60,6 +96,12 @@ public class UserController extends UserProfileController<CommonProfile> {
 
 	@Transactional
 	public Result update(String id) {
+		if(checkPrivilegesAdmin())
+		{
+			flash("Insufficient Privileges");
+			return ok(views.html.index.render());
+		}
+		
 		Form<User> userForm = form(User.class).bindFromRequest();
 		if(userForm.hasErrors()) {
 			return badRequest(edit.render(id, userForm));
@@ -68,5 +110,28 @@ public class UserController extends UserProfileController<CommonProfile> {
 			flash("success", "User " + id + " has been updated");
 			return ok(show.render(User.findById(id)));
 		}
+	}
+	
+	//This function returns false if current user does not possess any roles
+	public boolean checkPrivileges()
+	{
+		CommonProfile profile = getUserProfile();
+		//if you don't have admin role then redirect back to dashboard
+		if((User.findById(profile.getId()).role == null))
+		{
+			return false;
+		}
+		return true;
+	}
+	//This function returns false if current user does not possess the role Admin or SuperAdmin
+	public boolean checkPrivilegesAdmin()
+	{
+		CommonProfile profile = getUserProfile();
+		//if you don't have admin role then redirect back to dashboard
+		if(!(User.findById(profile.getId()).role != User.Role.Admin ||  User.findById(profile.getId()).role != User.Role.SuperAdmin))
+		{
+			return false;
+		}
+		return true;
 	}
 }
